+++ cmd/forklift/dev/plt/deployments.go
@@ -11,7 +11,10 @@ import (
 // ls-depl
 
 func lsDeplAction(c *cli.Context) error {
-	plt, caches, err := processFullBaseArgs(c, true, true)
+	plt, caches, err := processFullBaseArgs(c, processingOptions{
+		enableOverrides: true,
+		merge:           true,
+	})
 	if err != nil {
 		return err
 	}
@@ -22,7 +25,11 @@ func lsDeplAction(c *cli.Context) error {
 // show-depl
 
 func showDeplAction(c *cli.Context) error {
-	plt, caches, err := processFullBaseArgs(c, true, true)
+	plt, caches, err := processFullBaseArgs(c, processingOptions{
+		requireRepoCache: true,
+		enableOverrides:  true,
+		merge:            true,
+	})
 	if err != nil {
 		return err
 	}
@@ -34,7 +41,11 @@ func showDeplAction(c *cli.Context) error {
 // locate-depl-pkg
 
 func locateDeplPkgAction(c *cli.Context) error {
-	plt, caches, err := processFullBaseArgs(c, true, true)
+	plt, caches, err := processFullBaseArgs(c, processingOptions{
+		requireRepoCache: true,
+		enableOverrides:  true,
+		merge:            true,
+	})
 	if err != nil {
 		return err
 	}
@@ -47,7 +58,11 @@ func locateDeplPkgAction(c *cli.Context) error {
 
 func addDeplAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		plt, caches, err := processFullBaseArgs(c, true, true)
+		plt, caches, err := processFullBaseArgs(c, processingOptions{
+			requireRepoCache: true,
+			enableOverrides:  true,
+			merge:            true,
+		})
 		if err != nil {
 			return err
 		}
@@ -82,7 +97,11 @@ func addDeplAction(versions Versions) cli.ActionFunc {
 
 func rmDeplAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		plt, caches, err := processFullBaseArgs(c, false, false)
+		plt, caches, err := processFullBaseArgs(c, processingOptions{
+			requireRepoCache: true,
+			enableOverrides:  true,
+			merge:            true,
+		})
 		if err != nil {
 			return err
 		}
@@ -112,7 +131,11 @@ func rmDeplAction(versions Versions) cli.ActionFunc {
 
 func setDeplPkgAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		plt, caches, err := processFullBaseArgs(c, true, true)
+		plt, caches, err := processFullBaseArgs(c, processingOptions{
+			requireRepoCache: true,
+			enableOverrides:  true,
+			merge:            true,
+		})
 		if err != nil {
 			return err
 		}
@@ -144,7 +167,11 @@ func setDeplPkgAction(versions Versions) cli.ActionFunc {
 
 func addDeplFeatAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		plt, caches, err := processFullBaseArgs(c, true, true)
+		plt, caches, err := processFullBaseArgs(c, processingOptions{
+			requireRepoCache: true,
+			enableOverrides:  true,
+			merge:            true,
+		})
 		if err != nil {
 			return err
 		}
@@ -178,7 +205,11 @@ func addDeplFeatAction(versions Versions) cli.ActionFunc {
 
 func rmDeplFeatAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		plt, caches, err := processFullBaseArgs(c, true, true)
+		plt, caches, err := processFullBaseArgs(c, processingOptions{
+			requireRepoCache: true,
+			enableOverrides:  true,
+			merge:            true,
+		})
 		if err != nil {
 			return err
 		}
@@ -210,7 +241,11 @@ func rmDeplFeatAction(versions Versions) cli.ActionFunc {
 
 func setDeplDisabledAction(versions Versions, setting bool) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		plt, caches, err := processFullBaseArgs(c, true, true)
+		plt, caches, err := processFullBaseArgs(c, processingOptions{
+			requireRepoCache: true,
+			enableOverrides:  true,
+			merge:            true,
+		})
 		if err != nil {
 			return err
 		}