+++ cmd/forklift/dev/plt/downloads.go
@@ -12,7 +12,11 @@ import (
 
 func cacheDlAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		plt, caches, err := processFullBaseArgs(c, true, true)
+		plt, caches, err := processFullBaseArgs(c, processingOptions{
+			requireRepoCache: true,
+			enableOverrides:  true,
+			merge:            true,
+		})
 		if err != nil {
 			return err
 		}