+++ cmd/forklift/dev/plt/images.go
@@ -11,7 +11,11 @@ import (
 // ls-img
 
 func lsImgAction(c *cli.Context) error {
-	plt, caches, err := processFullBaseArgs(c, true, true)
+	plt, caches, err := processFullBaseArgs(c, processingOptions{
+		requireRepoCache: true,
+		enableOverrides:  true,
+		merge:            true,
+	})
 	if err != nil {
 		return err
 	}
@@ -30,7 +34,11 @@ func lsImgAction(c *cli.Context) error {
 
 func cacheImgAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		plt, caches, err := processFullBaseArgs(c, true, true)
+		plt, caches, err := processFullBaseArgs(c, processingOptions{
+			requireRepoCache: true,
+			enableOverrides:  true,
+			merge:            true,
+		})
 		if err != nil {
 			return err
 		}