+++ internal/app/forklift/pallets-imports.go
@@ -2,6 +2,8 @@ package forklift
 
 import (
 	"io/fs"
+	"path"
+	"slices"
 	"strings"
 
 	"github.com/bmatcuk/doublestar/v4"
@@ -57,6 +59,112 @@ func ResolveImports(
 	}
 
 	return resolvedImports, nil
+}
+
+// Evaluate returns a list of target file paths and a mapping between target file paths and source
+// file paths relative to the attached pallet's FS member. Directories are excluded from this
+// mapping.
+func (i *ResolvedImport) Evaluate() (map[string]string, error) {
+	pathMappings := make(map[string]string) // target -> source
+	for _, modifier := range i.Def.Modifiers {
+		switch modifier.Type {
+		default:
+			return pathMappings, errors.Errorf("unknown modifier type: %s", modifier.Type)
+		case ImportModifierTypeAdd:
+			if err := applyAddModifier(modifier, i.Pallet, pathMappings); err != nil {
+				return pathMappings, err
+			}
+		case ImportModifierTypeRemove:
+			if err := applyRemoveModifier(modifier, pathMappings); err != nil {
+				return pathMappings, err
+			}
+		}
+	}
+	return pathMappings, nil
+}
+
+func applyAddModifier(
+	modifier ImportModifier, pallet *FSPallet, pathMappings map[string]string,
+) error {
+	for _, matcher := range modifier.OnlyMatchingAny {
+		sourcePattern := modifier.Source
+		if matcher != "" {
+			sourcePattern = path.Join(sourcePattern, matcher)
+		}
+		sourcePattern = strings.TrimPrefix(sourcePattern, "/")
+		sourceFiles, err := globWithChildren(pallet.FS, sourcePattern, doublestar.WithFilesOnly())
+		if err != nil {
+			return err
+		}
+		for _, sourcePath := range sourceFiles {
+			targetPath := path.Join("/", modifier.Target, strings.TrimPrefix(
+				path.Join("/", sourcePath), path.Join("/", modifier.Source),
+			))
+			pathMappings[targetPath] = path.Join("/", sourcePath)
+		}
+	}
+	return nil
+}
+
+func globWithChildren(
+	fsys core.PathedFS, pattern string, opts ...doublestar.GlobOption,
+) ([]string, error) {
+	fileMatches, err := doublestar.Glob(fsys, pattern, opts...)
+	if err != nil {
+		return nil, errors.Wrapf(
+			err, "couldn't search for files in %s matching pattern %s", fsys.Path(), pattern,
+		)
+	}
+	pattern = path.Join(pattern, "**")
+	childMatches, err := doublestar.Glob(fsys, pattern, opts...)
+	if err != nil {
+		return nil, errors.Wrapf(
+			err, "couldn't search for files in %s matching pattern %s", fsys.Path(), pattern,
+		)
+	}
+	return slices.Concat(fileMatches, childMatches), nil
+}
+
+func applyRemoveModifier(modifier ImportModifier, pathMappings map[string]string) error {
+	for _, matcher := range modifier.OnlyMatchingAny {
+		if matcher == "" {
+			matcher = "**"
+		} else {
+			matcher = path.Join(matcher, "**")
+		}
+		targetPattern := path.Join(modifier.Target, matcher)
+		for target := range pathMappings {
+			matched, err := matchWithChildren(targetPattern, target)
+			if err != nil {
+				return err
+			}
+			if !matched {
+				continue
+			}
+			delete(pathMappings, target)
+		}
+	}
+	return nil
+}
+
+func matchWithChildren(pattern, name string) (bool, error) {
+	baseMatches, err := doublestar.Match(pattern, name)
+	if err != nil {
+		return false, errors.Wrapf(
+			err, "couldn't check whether %s matches pattern %s", name, pattern,
+		)
+	}
+	if baseMatches {
+		return true, nil
+	}
+	pattern = path.Join(pattern, "**")
+	childMatches, err := doublestar.Match(pattern, name)
+	if err != nil {
+		return false, errors.Wrapf(
+			err, "couldn't check whether %s matches pattern %s", name, pattern,
+		)
+	}
+	return childMatches, nil
 }
 
 // Import