+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (cli) Added a `[dev] plt ls-imp` command to list all file import groups declared by the local/development pallet.
 - (cli) Added a `[dev] plt show-imp` command to show the specified file import group declared by the local/development pallet.
 - (cli) Now the `[dev] plt cache-all` command, and all commands which can cache staging requirements, will cache pallets required by the local/development pallet.
+- (cli) Now all `[dev] plt` commands are evaluated on the merged pallet (i.e. with file imports) if the pallet imports files from other pallets.
 - (cli) Now the `[dev] plt show-pkg` and `cache show-pkg` commands also print information about file exports.
 - (cli) Added an optional `--plts` flag to `dev plt` for overriding version-locked required pallets with pallets from other directories, like the existing `--repos` flag.
 
@@ -25,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (Breaking change; cli) Now enabled feature flags are ordered alphabetically when sequencing file export operations declared by feature flags.
 - (cli) Now the commands for viewing a pallet/repo (e.g. `[dev] plt show`) truncate the printout of the pallet/repo's readme file to the first ten lines of the file, to prevent long readme files from clogging up the command output.
 - (cli) Now the `[dev] plt rm-repo` command only deletes the version lock file for the specified repository, instead of deleting the entire subdirectory for the repository.
+- (cli) Now a pallet can include deployments for local packages even if it's missing a `forklift-repository.yml` file in the pallet root - in such cases, the repository declaration is automatically inferred from the `forklift-pallet.yml` file for the purposes of using the pallet. However, a `forklift-repository.yml` file is still needed to make the pallet usable as a Forklift package repository by other pallets.
 
 ### Fixed